+++ config.go
@@ -5,6 +5,7 @@ import (
 	"runtime"
 
 	"github.com/gnames/gnfmt"
+	"github.com/gnames/gnparser/ent/nomcode"
 )
 
 // Config keeps settings that might affect how parsing is done,
@@ -12,6 +13,11 @@ import (
 type Config struct {
 	// BatchSize sets the maximum number of elements in names-strings slice.
 	BatchSize int
+
+	// Code contains optional nomenclatural code value. This option is
+	// useful to solve ambiguous parsing cases and to add cultivar botanical
+	// rules.
+	nomcode.Code
 
 	// Debug sets a "debug" state for parsing. The debug state forces output
 	// format to showing parsed ast tree.
@@ -42,10 +48,6 @@ type Config struct {
 	// WithCapitalization flag, when true, the first letter of a name-string
 	// is capitalized, if appropriate.
 	WithCapitalization bool
-
-	// WithCultivars flag, when true, cultivar names will be parsed and
-	// modify cardinality, normalized and canonical output.
-	WithCultivars bool
 
 	// WithDetails can be set to true when a simplified output is not sufficient
 	// for obtaining a required information.
@@ -93,17 +95,10 @@ func OptDebug(b bool) Option {
 	}
 }
 
-// OptFormat takes a string (one of 'csv', 'compact', 'pretty') to set
-// the formatting option for the CLI or Web presentation. If some other
-// string is entered, the default, 'CSV' format is set, accompanied by a
-// warning.
-func OptFormat(s string) Option {
+// OptFormat sets the formatting option for CLI or Web presentation.
+// It accepts a gnfmt.Format value to control the output format.
+func OptFormat(f gnfmt.Format) Option {
 	return func(cfg *Config) {
-		f, err := gnfmt.NewFormat(s)
-		if err != nil {
-			f = gnfmt.CSV
-			slog.Warn("Set default CSV format due to error", "error", err)
-		}
 		cfg.Format = f
 	}
 }
@@ -145,10 +140,10 @@ func OptWithCapitaliation(b bool) Option {
 	}
 }
 
-// OptWithCultivars sets the EnableCultivars field.
-func OptWithCultivars(b bool) Option {
+// OptCode sets Code field
+func OptCode(c nomcode.Code) Option {
 	return func(cfg *Config) {
-		cfg.WithCultivars = b
+		cfg.Code = c
 	}
 }
 
@@ -203,6 +198,7 @@ func NewConfig(opts ...Option) Config {
 		BatchSize:      50_000,
 		IgnoreHTMLTags: false,
 		Port:           8080,
+		Code:           nomcode.Unknown,
 	}
 	for i := range opts {
 		opts[i](&cfg)