+++ gnparser/cmd/root.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/dustin/go-humanize"
+	"github.com/gnames/gnfmt"
 	"github.com/gnames/gnparser"
 	"github.com/gnames/gnparser/ent/parsed"
 	"github.com/gnames/gnparser/io/web"
@@ -77,6 +78,8 @@ gnparser -j 5 -p 8080
 		withNoOrderFlag(cmd)
 		withCapitalizeFlag(cmd)
 		withEnableCultivarsFlag(cmd)
+		// overrides Cultivar flag
+		codeFlag(cmd)
 		withPreserveDiaeresesFlag(cmd)
 		batchSizeFlag(cmd)
 		spGrCutFlag(cmd)
@@ -94,7 +97,7 @@ gnparser -j 5 -p 8080
 			slog.SetDefault(logger)
 
 			webopts := []gnparser.Option{
-				gnparser.OptFormat("compact"),
+				gnparser.OptFormat(gnfmt.CompactJSON),
 				gnparser.OptWithWebLogs(withWebLogsFlag(cmd)),
 			}
 			cfg = gnparser.NewConfig(webopts...)
@@ -119,7 +122,6 @@ gnparser -j 5 -p 8080
 			debugName(data, cfg)
 			os.Exit(0)
 		}
-
 		parse(data, cfg)
 	},
 }
@@ -139,7 +141,18 @@ func init() {
 		"maximum number of names in a batch send for processing.")
 
 	rootCmd.Flags().BoolP("cultivar", "C", false,
-		"include cultivar epithets and graft-chimeras in normalized and canonical outputs")
+		"parse according to  cultivar code ICNCP (DEPRECATED, use nomenclatural-code instead)")
+
+	codeHelp := `Modifies the parser's behavior in ambiguous cases, sometimes 
+introducing additional parsing rules.
+
+Accepted values are:
+  - 'bot', 'icn', 'botanical' for botanical code
+  - 'cult', 'icncp', 'cultivar' for cultivar code
+  - 'zoo', 'iczn', 'zoological' for zoological code
+
+If not set, the parser will attempt to determine the appropriate code/s.`
+	rootCmd.Flags().StringP("nomenclatural-code", "n", "", codeHelp)
 
 	rootCmd.Flags().BoolP("capitalize", "c", false,
 		"capitalize the first letter of input name-strings")
@@ -149,9 +162,15 @@ func init() {
 
 	rootCmd.Flags().BoolP("details", "d", false, "provides more details")
 
-	formatHelp := "sets output format. Can be one of:\n" +
-		"'csv', 'tsv', 'compact', 'pretty'\n" +
-		"default is 'csv'"
+	formatHelp := `Sets the output format.
+
+Accepted values are:
+  - 'csv': Comma-separated values
+  - 'tsv': Tab-separated values
+  - 'compact': Compact JSON format
+  - 'pretty': Human-readable JSON format
+
+If not set, the output format defaults to 'csv'.`
 	rootCmd.Flags().StringP("format", "f", "", formatHelp)
 
 	rootCmd.Flags().BoolP("ignore_tags", "i", false,